# preppulse/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signUp
      tags: [auth]
      summary: Create an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Email already registered
  /auth/login:
    post:
      operationId: logIn
      tags: [auth]
      summary: Sign in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get the signed-in user and profile
      responses:
        "200":
          description: Current session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Not signed in
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request a password recovery email
      description: Always returns success for a well-formed email to avoid leaking account existence.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: Recovery email sent (or account not found)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "400":
          description: Invalid email
        "503":
          description: Recovery service unavailable
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Set a new password using a recovery token
      description: Verifies the recovery access token, provisions the account if needed and signs the user in.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password set and session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid or expired token, or weak password
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Set or change the signed-in user's password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "400":
          description: Current password is incorrect
        "401":
          description: Not signed in
...
components:
  schemas:
...
    AuthUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        handle:
          type: string
        hasPassword:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, email, handle, createdAt]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/AuthUser"
        profile:
          $ref: "#/components/schemas/Profile"
      required: [user, profile]
    SignupInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 8
        handle:
          type: string
          minLength: 2
      required: [email, password]
    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    ForgotPasswordInput:
      type: object
      properties:
        email:
          type: string
      required: [email]
    ResetPasswordInput:
      type: object
      properties:
        accessToken:
          type: string
        newPassword:
          type: string
          minLength: 8
      required: [accessToken, newPassword]
    ChangePasswordInput:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 8
      required: [newPassword]
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]
...
    Profile:
      type: object
      properties:
        handle:
          type: string
        email:
          type: string
        avatarUrl:
          type: string
          nullable: true
        examTrack:
          type: string
          enum: [jee_main, neet]
        stage:
          type: string
          enum: [class_11, class_12, dropper]
        targetYear:
          type: number
        examDate:
          type: string
          format: date
          nullable: true
        dailyGoalMinutes:
          type: number
        weeklyGoalMinutes:
          type: number
        focusMode:
          type: boolean
        showOnLeaderboard:
          type: boolean
        isAdmin:
          type: boolean
        profileCode:
          type: string
      required: [handle, email, avatarUrl, examTrack, stage, targetYear, examDate, dailyGoalMinutes, weeklyGoalMinutes, focusMode, showOnLeaderboard, isAdmin, profileCode]
...